api/requirements.txt
beautifulsoup4==4.8.1
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
lxml==4.4.2
numpy==1.17.4
pandas==0.25.3
requests==2.22.0
